Reverse changed catgeory add and course add. It was reversed before.
